refactor: Update AGI library and API
Replaced the 'agi' and 'asterisk.io' libraries with 'agi-async'. This change modernizes the AGI handling by using a more up-to-date and potentially better-maintained library. The AGI server initialization and handler functions have been updated to reflect the new library's API, specifically changing the context object to a channel object and updating method calls like `createServer` to `new AGI` and `listen` to an event listener.